Danny, please review.
QueryTraversal.java is preliminary - you and I need to talk about exception handling.  But the basic logic is there, and I think it helps us all understand how things are going to work to get it checked in.

QueryTraversalTest.java is a pretty good start.  I feel better about the test spi implementation to see that it works (basically) with all those different batch sizes.