Danny, please review.
First - my bad - I shouldn't have included MockInstantiator.java in R115.  It should have gone in this checkin.  So please review MickInstantiator from R115 with this.

MockInstantiator.java - reorganized to provide support for the new API's for getting the Auth*Managers.
SpringConnectorInstantiator.java - removed an embedded subclass and made it a class in its own right, so MockInstantiator could use it.
ConnectorInterfaces.java - the new class