Issue 4524060: Change SnapshotReader for backward compatibility
In 2.6.0 we used to start the record in snapshot file as follows 
----
{"SCANTIME":1305745733104,"MODTIME":1266561643327,"PATH":"smb://w2k8-iis/share001/doc/Cygwin.lnk","ACL":{},"STABLE":false,"CHECKSUM":"e54cf0cebd8f6a0fe67e276a40e6523dac7d95b2","FILESYS":"smb"}
----

After the re-architecture the record in snapshot file is as follows
----
192#{"SCANTIME":1305745733104,"MODTIME":1266561643327,"PATH":"smb://w2k8-iis/share001/doc/Cygwin.lnk","ACL":{},"STABLE":false,"CHECKSUM":"e54cf0cebd8f6a0fe67e276a40e6523dac7d95b2","FILESYS":"smb"}
----

The new version of SnapshotReader is not backward compatible so after upgrade it
won't work with 2.6.0 connectors 

This CL makes this behaviour compatible.

Changes done:
1. In the constructor, check if the snapshot is old version or new. Depending on
that it instantiates appropriate RecordReader
2. Created New i/f : RecordReader and 2 impl : a. LineBasedRecordReader b.
LengthBasedRecordReader
3. Based on the version of the snapshot file: one of the impls is assigned
//TODO: Evaluate whether we can introduce version number in the snapshot file
and depending on that designate the record reader. 
All test cases passed. Added a new test case to check backward compatibility. 

Issue URL: http://codereview.appspot.com/4524060/