Issue 4581062: Introduce a configurable 'introduceDelayAfterEachScan' parameter to ChangeQueue
There have been many cases where customers feel that the load on the repository
should be reduced (for LDAP connector). Going ahead, this will be also an issue
with DB connector which is also now based on diffing.

We have plans to introduce schedule in diffing in future but minimally we should
expose an option to add delay (sleep time) after every scan as opposed to the
current scenario where we add delay after scans with no changes found.

This CL implements the configurable 'introduceDelayAfterEachScan' parameter and
implements the above mentioned feature. 

CL URL: http://codereview.appspot.com/4581062/