GetDocumentContent servlet logs ConnectorNotFoundException and RepositoryDocumentExceptions less verbosely.
The most likely cause of ConnectorNotFoundException in this case
would be the user deleted a connector.  The GSA doesn't know about
it until it trys to crawl the documents and gets a 404.  This did
not need to be a Warning and certainly didn't need a stack trace.

Similarly, the most likely cause of a RepositoryDocumentExeption
here is the document was deleted.

Change Log:
-----------
M projects/connector-manager/source/java/com/google/enterprise/connector/servlet/GetDocumentContent.java

Code Review:
------------
URL: http://codereview.appspot.com/5689077