CM Fix Bug 6892118: GetDocumentContent sets "Content-Type" Header.
The GetDocumentContent servlet now sets the Content-Type header
if the document metadata supplies a "google:mimetype" property.
If there no mime type is supplied by the connector, the servlet
returns a Content-Type of "text/html" (SpiConstants.DEFAULT_MIMETYPE).
This is done to be consistent with the behaviour of XmlFeed, even
though the GSA would actually prefer to get "application/octet-stream"
in this case.

Code Review:
http://codereview.appspot.com/6455063